Remove SIGINT handler.
Since the signal handler doesn't do any cleanup, it is not needed and
only causes a crash on exit. Even if we wanted it I don't see the
point of registering it in a separate thread unless you use that
thread to do some cleanup while the other threads are still running.
By default SIGINT should already terminate the program, which is all
we currently want.